+++ app/src/test/kotlin/com/wespot/vote/fixture/BallotFixture.kt
