+++ app/src/test/kotlin/com/wespot/vote/fixture/BallotJpaEntityFixture.kt
