+++ app/src/test/kotlin/com/wespot/vote/fixture/VoteFixture.kt
