+++ app/src/test/kotlin/com/wespot/vote/fixture/VoteJpaEntityFixture.kt
