+++ app/src/test/kotlin/com/wespot/voteoption/fixture/VoteOptionFixture.kt
