+++ app/src/test/kotlin/com/wespot/voteoption/fixture/VoteOptionJpaEntityFixture.kt
@@ -1,0 +1,15 @@
+package com.wespot.voteoption.fixture
+
+import com.wespot.common.BaseEntity
+import com.wespot.voteoption.VoteOptionJpaEntity
+import java.time.LocalDateTime
+
+object VoteOptionJpaEntityFixture {
+
+    fun createMock() = VoteOptionJpaEntity(
+        id = 1L,
+        content = "Mock 질문을 만듭니다.",
+        baseEntity = BaseEntity(LocalDateTime.now(), null),
+    )
+
+}
